+++ earth2studio/models/dx/solarradiation_afno.py
@@ -363,7 +363,8 @@ class SolarRadiationAFNO1H(SolarRadiationAFNO):
 
     Badges
     ------
-    region:global class:mrf product:solar year:2024 gpu:40gb
+    region:global class:medium-range product:solar year:2024 gpu:40gb
+    provider:nvidia backend:pytorch
     """
 
     freq = "1h"
@@ -387,7 +388,8 @@ class SolarRadiationAFNO6H(SolarRadiationAFNO):
 
     Badges
     ------
-    region:global class:mrf product:solar year:2024 gpu:40gb
+    region:global class:medium-range product:solar year:2024 gpu:40gb
+    provider:nvidia backend:pytorch
     """
 
     freq = "6h"